fix: index.js - eval suite if it is json string.
If prototype.js is specified in the karma config, the specSucess fails.
This is because the result.suite is a json string instead of the
array which it evalutes into.  This patch resolves the issue by
eval'ing the json back into an array.  I suppose it would be a problem
if the name of the suite began with "[", however.  Not sure how
likely that is?

Closes #30