Release memory held by a testsuite after we're done with it.
When a test finishes and the result has been written out the reference to the testsuite in the suites object can be dropped.
Ideally we'd drop the reference from the array entirely but that involves recreating the array and *that* might race.
So instead we just make the testsuite object point to 'null'.