fix: allow special characters and emoji in output
If an emoji is used anywhere in either suite description or inside an
error from a test (e.g. "Expected '👎' to be '👍'"), the reporter
chrashes with "Invalid character (�) in string: 😄 at index 0".

The root cause of this is xmlbuilder, that uses a version that is
3 years old at this time. The current version does not throw any
errors for this.

This PR will suggest updating to the latest version of xmlbuilder,
that includes a lot of bugfixes. As far as I could determine,
there were no breaking changes for this package.